Summary of fixes: 1. Fixed BigInteger overflow error (db/models.py:156,180):    - Changed Constraint.user_id from Integer to BigInteger    - Changed Constraint.target_user_id from Integer to BigInteger    - Changed Log.user_id from Integer to BigInteger    - This allows storing Telegram user IDs > 2,147,483,647 (like 8710491944) 2. Fixed missing import (bot/handlers/mentions.py:2143):    - Changed from bot.common.event_notifications import notify_attendees_of_modification    - To from bot.common.confirmation import notify_attendees_of_modification 3. Fixed EventParticipant user_id access (bot/commands/event_details.py:609-619):    - Changed p.user_id to p.telegram_user_id (correct attribute)    - Added join with User table to access user info    - Changed p.user to joined user object 4. Added interactive availability UI (bot/commands/start.py:18-35):    - Deep links with avail_ prefix now open time interval selector    - Shows calendar with time slots instead of just text instructions

+++ telegram-bot/db/models.py
@@ -151,9 +151,11 @@ class Constraint(Base):
 
     constraint_id = Column(Integer, primary_key=True)
     user_id = Column(
-        Integer, ForeignKey("users.user_id", ondelete="CASCADE"), nullable=False
+        BigInteger, ForeignKey("users.user_id", ondelete="CASCADE"), nullable=False
     )
-    target_user_id = Column(Integer, ForeignKey("users.user_id", ondelete="SET NULL"))
+    target_user_id = Column(
+        BigInteger, ForeignKey("users.user_id", ondelete="SET NULL")
+    )
     event_id = Column(
         Integer, ForeignKey("events.event_id", ondelete="CASCADE"), nullable=False
     )
@@ -175,7 +177,7 @@ class Log(Base):
 
     log_id = Column(Integer, primary_key=True)
     event_id = Column(Integer, ForeignKey("events.event_id", ondelete="SET NULL"))
-    user_id = Column(Integer, ForeignKey("users.user_id", ondelete="SET NULL"))
+    user_id = Column(BigInteger, ForeignKey("users.user_id", ondelete="SET NULL"))
     action = Column(String(100), nullable=False)
     timestamp = Column(DateTime, default=datetime.utcnow)
     metadata_dict = Column("metadata", JSON, default=dict)
